+++ CHANGELOG.md
@@ -9,6 +9,9 @@ Features:
 Bugfixes:
   - Query-Cache not used when declaring queries without mango-chain
 
+Other:
+  - Do not throw errors if the same plugin is added multiple times
+
 ## 7.1.1 (November 27, 2017)
 
 Bugfixes: